sentiment_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
transformers==4.41.0
torch==2.3.0
emoji==2.11.1
nltk==3.8.1
alembic==1.13.1
aiosqlite==0.20.0
slowapi==0.1.9
python-dotenv==1.0.1
httpx==0.27.0